fix(phase-2): adclib %zu format spec + adc.lua salt-length cap (CI fix)
CI smoke on Linux failed with "invalid option '%z' to 'lua_pushfstring'".
Two underlying bugs, both pre-existing:

1. adclib/adclib.cpp:217 + :246 use luaL_error with the printf
   "%zu" size_t format specifier. Lua 5.4's lua_pushfstring only
   supports %s/%d/%c/%I/%f/%p/%q/%U/%%. Hit the moment hashpas is
   called with an out-of-range salt and the error path tries to
   format saltBytes - the format-string parser sees "%z", errors
   out, and propagates "invalid option '%z'" instead of the real
   "salt length X out of range" message. Fix: cast to int and use
   "%d" (saltBytes is bounded by MAX_SALT_BYTES + a small overshoot,
   fits in int). Both hash_pas and hash_pas_oldschool get the same
   fix.

2. core/adc.lua:13 builds the salt-arg to adclib.hashpas as
   "os.date() .. os.clock() .. os.time() .. fixed-60-char-salt".
   adclib.hashpas requires the second arg (salt) to produce <= 64
   bytes of base32-decoded data, i.e. <= 102 chars of base32 string.
   On Windows os.date() returns ~24 chars (just under the limit);
   on Linux with C.UTF-8 locale it can return 30+ chars (pushes the
   total over 102 and triggers the out-of-range error). Phase 1
   shipped on Windows and never noticed; Phase 2 CI on ubuntu-
   latest surfaces it.

   Fix: cap the variable component (os.date+clock+time) at 30 chars
   via string.sub(1, 30) so the total salt length is always within
   the bounds adclib enforces. Generated PID per-install changes vs
   the pre-fix Windows behaviour ONLY for installs with no existing
   cfg/id.lua (i.e. fresh) - existing installs read the cached
   id.lua and don't re-call createid. Acceptable for a Phase 2 fix.

Verification (local Windows smoke from build/install/announcer/):
  ./lua.exe -e 'dofile "core/init.lua"' -> "SMOKE OK: bootstrap
  reaches ssl.newcontext" (expected - no cert installed).

The %zu format-spec bug exists in the parent luadch repo's vendored
adclib too. Hub doesn't trigger it because hub.lua calls hashpas
with operator-supplied salts (always short). Announcer triggers it
because adc.createid programmatically constructs a long salt at
PID-generation time. Worth a follow-up sync back to hub.

+++ adclib/adclib.cpp
@@ -215,8 +215,11 @@ int hash_pas(lua_State* L)
 
     size_t saltBytes = salt_len * 5 / 8;
     if (saltBytes == 0 || saltBytes > MAX_SALT_BYTES) {
-        return luaL_error(L, "hashpas: salt length %zu out of range",
-                          saltBytes);
+        // %zu is NOT supported by lua_pushfstring in Lua 5.4 (only %s
+        // %d %c %I %f %p %q %U %%). Cast to int + use %d. saltBytes
+        // <= MAX_SALT_BYTES + a small overshoot, fits in int.
+        return luaL_error(L, "hashpas: salt length %d out of range",
+                          (int)saltBytes);
     }
     unsigned char chunk[MAX_SALT_BYTES];
 
@@ -240,8 +243,9 @@ int hash_pas_oldschool(lua_State* L)
 
     size_t saltBytes = salt_len * 5 / 8;
     if (saltBytes == 0 || saltBytes > MAX_SALT_BYTES) {
-        return luaL_error(L, "hasholdpas: salt length %zu out of range",
-                          saltBytes);
+        // Same Lua 5.4 lua_pushfstring %zu unsupported issue as hash_pas.
+        return luaL_error(L, "hasholdpas: salt length %d out of range",
+                          (int)saltBytes);
     }
     unsigned char chunk1[MAX_SALT_BYTES];
     unsigned char chunk2[SIZE];
